+++ src/main/java/com/tinDev/repository/vacancy/VacancyRepository.java
@@ -3,5 +3,5 @@ package com.tinDev.repository.vacancy;
 import com.tinDev.models.vanancy.Vacancy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VacancyRepository extends JpaRepository<Vacancy, Integer> {
+public interface VacancyRepository extends JpaRepository<Vacancy, Long> {
 }
