+++ src/main/java/com/tinDev/services/user/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UserService {
-    Optional<User> findById(int id);
+    Optional<User> findById(Long id);
     List<User> findAll();
     User save(User user);
-    void deleteById(int id);
+    void deleteById(long id);
 }
